Replaces UnicodeUtils.display_width with Unicode::DisplayWidth.of
This is a performance improvement to support the tty-table project.
Currently the tty-table gem loads very slowly, which means it can't
really be used for command line applications. One major source of
the slowness is loading the "unicode/display_width" file.

To speed it up, I've replaced `UnicodeUtils.display_width` with
`Unicode::DisplayWidth.of`, which loads much more quickly.